Make some organizational improvements
