+++ src/components/FiltersDrawer/FilterGroup/FilterGroup.test.tsx
@@ -1,6 +1,6 @@
 import { CaseSensitivity, MatchType } from "constants/enums";
 import { render, screen, userEvent } from "test_utils";
-import Filter from ".";
+import FilterGroup from ".";
 
 const defaultFilter = {
   caseSensitive: CaseSensitivity.Insensitive,
@@ -14,7 +14,7 @@ describe("filters", () => {
 
   it("should display the name of the filter", () => {
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={jest.fn()}
         editFilter={jest.fn()}
         filter={defaultFilter}
@@ -25,7 +25,7 @@ describe("filters", () => {
 
   it("should be able to toggle editing", async () => {
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={jest.fn()}
         editFilter={jest.fn()}
         filter={defaultFilter}
@@ -50,7 +50,7 @@ describe("filters", () => {
   it("should call editFilter with the correct parameters", async () => {
     const editFilter = jest.fn();
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={jest.fn()}
         editFilter={editFilter}
         filter={defaultFilter}
@@ -73,7 +73,7 @@ describe("filters", () => {
   it("should toggle between visibility icons when they are clicked", async () => {
     const editFilter = jest.fn();
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={jest.fn()}
         editFilter={editFilter}
         filter={defaultFilter}
@@ -88,7 +88,7 @@ describe("filters", () => {
   it("should call deleteFilter with the correct parameters", async () => {
     const deleteFilter = jest.fn();
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={deleteFilter}
         editFilter={jest.fn()}
         filter={defaultFilter}
@@ -102,7 +102,7 @@ describe("filters", () => {
   it("should be able to interact with Case Sensitivity segmented control", async () => {
     const editFilter = jest.fn();
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={jest.fn()}
         editFilter={editFilter}
         filter={defaultFilter}
@@ -134,7 +134,7 @@ describe("filters", () => {
   it("should be able to interact with Match Type segmented control", async () => {
     const editFilter = jest.fn();
     render(
-      <Filter
+      <FilterGroup
         deleteFilter={jest.fn()}
         editFilter={editFilter}
         filter={defaultFilter}
