+++ src/context/LogContext/index.tsx
@@ -12,7 +12,8 @@ import { FilterLogic, LogTypes } from "constants/enums";
 import { QueryParams } from "constants/queryParams";
 import { useFilterParam, useQueryParam } from "hooks/useQueryParam";
 import { ExpandedLines, ProcessedLogLines } from "types/logs";
-import { filterLogs, getMatchingLines } from "utils/filter";
+import filterLogs from "utils/filterLogs";
+import { getMatchingLines } from "utils/matchingLines";
 import { getColorMapping } from "utils/resmoke";
 import searchLogs from "utils/searchLogs";
 import useLogState from "./state";
