+++ src/types/logs.ts
@@ -1,12 +1,24 @@
+import { CaseSensitivity, MatchType } from "constants/enums";
+
 type ExpandedLine = [number, number];
 type ExpandedLines = ExpandedLine[];
 
 type ProcessedLogLine = number | number[];
 type ProcessedLogLines = ProcessedLogLine[];
 
+type Filter = {
+  name: string;
+  visible: boolean;
+  caseSensitive: CaseSensitivity;
+  matchType: MatchType;
+};
+type Filters = Filter[];
+
 export type {
   ExpandedLine,
   ExpandedLines,
+  Filter,
+  Filters,
   ProcessedLogLine,
   ProcessedLogLines,
 };
