+++ src/utils/filterLogs/index.ts
@@ -1,5 +1,3 @@
-import { CaseSensitivity, FilterLogic, MatchType } from "constants/enums";
-import type { ParsedFilter } from "types/filters";
 import { ExpandedLines, ProcessedLogLines } from "types/logs";
 import { isCollapsedRow } from "utils/collapsedRow";
 import { isExpanded } from "utils/expandedLines";
@@ -24,7 +22,7 @@ type FilterLogsParams = {
  * @returns an array of numbers that indicates which log lines should be displayed, and which log lines
  * should be collapsed
  */
-export const filterLogs = ({
+const filterLogs = ({
   logLines,
   matchingLines,
   bookmarks,
@@ -70,69 +68,4 @@ export const filterLogs = ({
   return filteredLines;
 };
 
-export const constructRegexToMatch = (visibleFilters: ParsedFilter[]) => {
-  const regexToMatch: { regex: RegExp; isMatch: boolean }[] = [];
-
-  visibleFilters.forEach((f) => {
-    const isMatch = f.matchType === MatchType.Exact;
-    if (f.caseSensitive === CaseSensitivity.Sensitive) {
-      regexToMatch.push({
-        regex: new RegExp(f.name),
-        isMatch,
-      });
-    } else {
-      regexToMatch.push({
-        regex: new RegExp(f.name, "i"),
-        isMatch,
-      });
-    }
-  });
-
-  return regexToMatch;
-};
-
-export const matchesFilters = (
-  line: string,
-  regexToMatch: { regex: RegExp; isMatch: boolean }[],
-  filterLogic: FilterLogic
-) => {
-  if (filterLogic === FilterLogic.And) {
-    return regexToMatch.every(({ regex, isMatch }) =>
-      isMatch ? regex.test(line) : !regex.test(line)
-    );
-  }
-  return regexToMatch.some(({ regex, isMatch }) =>
-    isMatch ? regex.test(line) : !regex.test(line)
-  );
-};
-
-/**
- * 'getMatchingLines' determines which log lines match the applied filters. This function returns undefined if no
- * filters have been applied.
- * @param logLines - list of strings representing the log lines
- * @param filters - list of filters being applied
- * @param filterLogic - specifies whether to use AND or OR filtering
- * @returns a set of numbers representing the matched lines, or undefined
- */
-export const getMatchingLines = (
-  logLines: string[],
-  filters: ParsedFilter[],
-  filterLogic: FilterLogic
-) => {
-  const set = new Set<number>();
-
-  const visibleFilters = filters.filter((f) => f.visible !== false);
-  if (visibleFilters.length === 0) {
-    return undefined;
-  }
-
-  const regexToMatch = constructRegexToMatch(filters);
-
-  logLines.forEach((line, idx) => {
-    if (matchesFilters(line, regexToMatch, filterLogic)) {
-      set.add(idx);
-    }
-  });
-
-  return set;
-};
+export default filterLogs;
